refactor: :fire: tax-mock-factory
