+++ apps/taxes/src/modules/avatax/order-fulfilled/avatax-order-fulfilled-payload-transformer.test.ts
@@ -1,14 +1,25 @@
+import { DocumentType } from "avatax/lib/enums/DocumentType";
+import { describe, expect, it } from "vitest";
 import { OrderFulfilledSubscriptionFragment } from "../../../../generated/graphql";
+import { AvataxConfig } from "../avatax-config";
+import { Payload } from "./avatax-order-fulfilled-adapter";
 import {
   AvataxOrderFulfilledPayloadTransformer,
   PROVIDER_ORDER_ID_KEY,
   getTransactionCodeFromMetadata,
 } from "./avatax-order-fulfilled-payload-transformer";
-import { describe, it, expect } from "vitest";
-import { DocumentType } from "avatax/lib/enums/DocumentType";
-import { Payload } from "./avatax-order-fulfilled-adapter";
-import { taxMockFactory } from "../../taxes/tax-mock-factory";
-import { avataxMockFactory } from "../avatax-mock-factory";
+
+// todo: add AvataxOrderFulfilledMockGenerator
+
+const MOCK_AVATAX_CONFIG: AvataxConfig = {
+  companyCode: "DEFAULT",
+  isAutocommit: false,
+  isSandbox: true,
+  name: "Avatax-1",
+  password: "password",
+  username: "username",
+  shippingTaxCode: "FR000000",
+};
 
 const MOCKED_METADATA: OrderFulfilledSubscriptionFragment["privateMetadata"] = [
   {
@@ -16,6 +27,89 @@ const MOCKED_METADATA: OrderFulfilledSubscriptionFragment["privateMetadata"] = [
     value: "transaction-code",
   },
 ];
+
+type OrderFulfilled = OrderFulfilledSubscriptionFragment;
+
+const ORDER_FULFILLED_MOCK: OrderFulfilled = {
+  id: "T3JkZXI6OTU4MDA5YjQtNDUxZC00NmQ1LThhMWUtMTRkMWRmYjFhNzI5",
+  created: "2023-04-11T11:03:09.304109+00:00",
+  privateMetadata: MOCKED_METADATA,
+  channel: {
+    id: "Q2hhbm5lbDoy",
+    slug: "channel-pln",
+  },
+  shippingAddress: {
+    streetAddress1: "123 Palm Grove Ln",
+    streetAddress2: "",
+    city: "LOS ANGELES",
+    countryArea: "CA",
+    postalCode: "90002",
+    country: {
+      code: "US",
+    },
+  },
+  billingAddress: {
+    streetAddress1: "123 Palm Grove Ln",
+    streetAddress2: "",
+    city: "LOS ANGELES",
+    countryArea: "CA",
+    postalCode: "90002",
+    country: {
+      code: "US",
+    },
+  },
+  total: {
+    net: {
+      amount: 183.33,
+    },
+    tax: {
+      amount: 12.83,
+    },
+  },
+  shippingPrice: {
+    net: {
+      amount: 48.33,
+    },
+  },
+  lines: [
+    {
+      productSku: "328223581",
+      productName: "Monospace Tee",
+      quantity: 1,
+      unitPrice: {
+        net: {
+          amount: 90,
+        },
+      },
+      totalPrice: {
+        net: {
+          amount: 90,
+        },
+        tax: {
+          amount: 8.55,
+        },
+      },
+    },
+    {
+      productSku: "328223580",
+      productName: "Polyspace Tee",
+      quantity: 1,
+      unitPrice: {
+        net: {
+          amount: 45,
+        },
+      },
+      totalPrice: {
+        net: {
+          amount: 45,
+        },
+        tax: {
+          amount: 4.28,
+        },
+      },
+    },
+  ],
+};
 
 describe("getTransactionCodeFromMetadata", () => {
   it("returns transaction code", () => {
@@ -30,8 +124,8 @@ describe("getTransactionCodeFromMetadata", () => {
 const transformer = new AvataxOrderFulfilledPayloadTransformer();
 
 const MOCKED_ORDER_FULFILLED_PAYLOAD: Payload = {
-  order: taxMockFactory.createOrderFulfilledMock({ privateMetadata: MOCKED_METADATA }),
-  config: avataxMockFactory.createMockAvataxConfig(),
+  order: ORDER_FULFILLED_MOCK,
+  config: MOCK_AVATAX_CONFIG,
 };
 
 describe("AvataxOrderFulfilledPayloadTransformer", () => {
